backend/requirements.txt
flask
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
python-dotenv
gunicorn
yfinance
requests
feedparser
